feat: close Realtime parity gaps
