+++ test/openai/logging_test.rb
@@ -1,113 +1,8 @@
 # frozen_string_literal: true
 
-require_relative "test_helper"
+require_relative "logging_test_case"
 
-class LoggingTest < Minitest::Test
-  extend Minitest::Serial
-
-  class CapturingLogger
-    attr_reader :events
-
-    def initialize
-      @events = []
-    end
-
-    def debug(message) = @events << [:debug, message]
-    def info(message) = @events << [:info, message]
-    def warn(message) = @events << [:warn, message]
-    def error(message) = @events << [:error, message]
-  end
-
-  class RaisingLogger
-    def debug(_message) = raise("debug logger failed")
-    def info(_message) = raise("info logger failed")
-    def warn(_message) = raise("warn logger failed")
-    def error(_message) = raise("error logger failed")
-  end
-
-  class StubHTTPClient < OpenAI::HTTPClient
-    def initialize(&execute)
-      super()
-      @execute = execute
-    end
-
-    def execute(request) = @execute.call(request)
-  end
-
-  class CloseableBody
-    include Enumerable
-
-    attr_reader :chunk_count, :close_count, :each_count
-
-    def initialize(*chunks)
-      @chunks = chunks
-      @chunk_count = 0
-      @close_count = 0
-      @each_count = 0
-    end
-
-    def close = (@close_count += 1)
-
-    def each
-      @each_count += 1
-      @chunks.each do |chunk|
-        @chunk_count += 1
-        yield(chunk)
-      end
-    end
-  end
-
-  class UnbufferableChunk < String
-    def byteslice(*) = raise("stream chunk was buffered")
-  end
-
-  class FailingBody
-    include Enumerable
-
-    attr_reader :close_count
-
-    def initialize(error)
-      @error = error
-      @close_count = 0
-    end
-
-    def each
-      yield("partial")
-      raise @error
-    end
-
-    def close = (@close_count += 1)
-  end
-
-  def setup
-    super
-    @openai_log = ENV.delete("OPENAI_LOG")
-  end
-
-  def teardown
-    @openai_log.nil? ? ENV.delete("OPENAI_LOG") : ENV["OPENAI_LOG"] = @openai_log
-    super
-  end
-
-  def diagnostic_client(
-    http_client: StubHTTPClient.new { raise "unexpected HTTP request" },
-    logger: nil,
-    log_level: nil,
-    on_retry: nil,
-    **options
-  )
-    transport = StubHTTPClient.new { |request| http_client.execute(request) }
-    client_options = {
-      api_key: "test-key",
-      **options,
-      http_client: transport,
-      logger: logger,
-      log_level: log_level,
-      on_retry: on_retry
-    }
-    OpenAI::Client.new(**client_options)
-  end
-
+class LoggingTest < OpenAILoggingTestCase
   def test_logging_is_off_by_default
     client = OpenAI::Client.new(api_key: "test-key")
 
@@ -756,72 +651,6 @@ class LoggingTest < Minitest::Test
     assert_equal(1, logger.events.count { |severity, _message| severity == :error })
     assert_empty(logger.events.select { |_severity, message| message.include?("request complete") })
     assert_equal(1, logger.events.count { |_severity, message| message.include?("response body") })
-  end
-
-  def test_raw_response_completion_is_logged_only_after_body_consumption
-    logger = CapturingLogger.new
-    source = CloseableBody.new("answer-sdp")
-    http_client = StubHTTPClient.new do |_request|
-      OpenAI::HTTPClient::Response.new(
-        status: 200,
-        headers: {"content-type" => "application/sdp", "x-request-id" => "req_raw"},
-        body: source
-      )
-    end
-    client = diagnostic_client(http_client: http_client, logger: logger, log_level: :info)
-
-    response = client.request_raw(method: :post, path: "realtime/calls")
-
-    assert_empty(logger.events)
-    assert_equal(["answer-sdp"], response.body.to_a)
-    assert_equal(1, source.close_count)
-    assert_equal([:info], logger.events.map(&:first))
-    assert_includes(logger.events.fetch(0).fetch(1), "request complete")
-  end
-
-  def test_raw_response_body_failure_logs_an_error_without_a_success_event
-    logger = CapturingLogger.new
-    failure = OpenAI::Errors::APIConnectionError.new(
-      url: URI("https://example.com/v1/realtime/calls")
-    )
-    source = FailingBody.new(failure)
-    http_client = StubHTTPClient.new do |_request|
-      OpenAI::HTTPClient::Response.new(
-        status: 200,
-        headers: {"content-type" => "application/sdp"},
-        body: source
-      )
-    end
-    client = diagnostic_client(http_client: http_client, logger: logger, log_level: :info)
-    response = client.request_raw(method: :post, path: "realtime/calls")
-
-    assert_empty(logger.events)
-    error = assert_raises(OpenAI::Errors::APIConnectionError) { response.body.to_a }
-
-    assert_same(failure, error)
-    assert_equal(1, source.close_count)
-    assert_equal([:error], logger.events.map(&:first))
-    assert_includes(logger.events.fetch(0).fetch(1), "request failed")
-    refute_includes(logger.events.fetch(0).fetch(1), "request complete")
-  end
-
-  def test_closing_a_partially_consumed_raw_response_closes_its_source
-    logger = CapturingLogger.new
-    source = CloseableBody.new("first", "second")
-    http_client = StubHTTPClient.new do |_request|
-      OpenAI::HTTPClient::Response.new(
-        status: 200,
-        headers: {"content-type" => "application/sdp"},
-        body: source
-      )
-    end
-    client = diagnostic_client(http_client: http_client, logger: logger, log_level: :info)
-    response = client.request_raw(method: :post, path: "realtime/calls")
-
-    assert_equal("first", response.body.first)
-
-    assert_equal(1, source.close_count)
-    assert_empty(logger.events)
   end
 
   def test_consumer_exceptions_are_not_logged_as_request_failures
